Updates for xmltree and parsexml modules
These are helpers and changes that would be nice if they
were in the xmltree and parsexml modules.

Some are existing procs that are not public, others are
general helpers.

No effort has been made yet to see if any of these changes
would be accepted upstream or not yet.